# Faster, cleaner pressure solvers for all topologies on CPUs and GPUs