--- a/src/Solvers/Solvers.jl
+++ b/src/Solvers/Solvers.jl
@@ -4,25 +4,29 @@ export
     BatchedTridiagonalSolver, solve_batched_tridiagonal_system!,
     PressureSolver, solve_for_pressure!
 
+using FFTW
 using CUDA
 using KernelAbstractions
 using KernelAbstractions.Extras.LoopInfo: @unroll
 
-using Oceananigans.Architectures: device, @hascuda, CPU, GPU, array_type
+using Oceananigans.Architectures: device, @hascuda, CPU, GPU, array_type, arch_array
 using Oceananigans.Utils
 using Oceananigans.Grids
+
 using Oceananigans.Grids: unpack_grid
 
-include("solver_utils.jl")
-include("batched_tridiagonal_solver.jl")
+"""
+    ω(M, k)
 
+Return the `M`th root of unity raised to the `k`th power.
+"""
+@inline ω(M, k) = exp(-2im*π*k/M)
+
+include("batched_tridiagonal_solver.jl")
 include("discrete_eigenvalues.jl")
 include("plan_transforms.jl")
 include("pressure_solver.jl")
-include("triply_periodic_pressure_solver.jl")
-include("horizontally_periodic_pressure_solver.jl")
-include("channel_pressure_solver.jl")
-include("box_pressure_solver.jl")
+include("solve_poisson_equation.jl")
 include("index_permutations.jl")
 include("solve_for_pressure.jl")