--- a/src/Solvers/Solvers.jl
+++ b/src/Solvers/Solvers.jl
@@ -22,9 +22,13 @@ Return the `M`th root of unity raised to the `k`th power.
 """
 @inline ω(M, k) = exp(-2im*π*k/M)
 
+reshaped_size(N, dim) = dim == 1 ? (N, 1, 1) :
+                        dim == 2 ? (1, N, 1) :
+                        dim == 3 ? (1, 1, N) : nothing
+
 include("batched_tridiagonal_solver.jl")
 include("poisson_eigenvalues.jl")
-include("plan_transforms.jl")
+include("transforms.jl")
 include("pressure_solver.jl")
 include("solve_poisson_equation.jl")
 include("index_permutations.jl")