--- a/src/Solvers/Solvers.jl
+++ b/src/Solvers/Solvers.jl
@@ -2,7 +2,7 @@ module Solvers
 
 export
     BatchedTridiagonalSolver, solve_batched_tridiagonal_system!,
-    PressureSolver, solve_for_pressure!
+    PressureSolver, solve_for_pressure!, solve_poisson_equation!
 
 using FFTW
 using CUDA