--- a/src/Solvers/Solvers.jl
+++ b/src/Solvers/Solvers.jl
@@ -2,7 +2,7 @@ module Solvers
 
 export
     BatchedTridiagonalSolver, solve_batched_tridiagonal_system!,
-    PressureSolver, solve_for_pressure!, solve_poisson_equation!
+    FFTBasedPoissonSolver, solve_for_pressure!, solve_poisson_equation!
 
 using FFTW
 using CUDA
@@ -30,7 +30,7 @@ include("batched_tridiagonal_solver.jl")
 include("poisson_eigenvalues.jl")
 include("discrete_transforms.jl")
 include("plan_transforms.jl")
-include("pressure_solver.jl")
+include("fft_based_poisson_solver.jl")
 include("solve_poisson_equation.jl")
 include("index_permutations.jl")
 include("solve_for_pressure.jl")