--- a/src/Solvers/solver_utils.jl
+++ b/src/Solvers/solver_utils.jl
@@ -1,6 +1,0 @@
-"""
-    ω(M, k)
-
-Return the `M`th root of unity raised to the `k`th power.
-"""
-@inline ω(M, k) = exp(-2im*π*k/M)