This PR is still a work-in-progress but I'm opening it to make the future design of the pressure solver module more transparent as we will be adding some new pressure solvers soon, including a conjugate-gradient solver by @christophernhill.

# Motivation

In PR #290 I implemented a pressure solver for the `(Periodic, Bounded, Bounded)` channel topology using the 2D fast cosine transform algorithm described by Makhoul (1982) as CUFFT does not provide cosine transforms for the GPU and does not support FFTs along non-batched dimensions (see https://github.com/JuliaGPU/CUDA.jl/issues/119).

This has been a very unpopular decision for good reasons. The 2D DCT algorithm is quite slow (channels are ~2x slower than doubly periodic on GPUs) and is quite complicated. Due to my inexperience, I didn't realize that transposing the array to do the FFT was the way forward.

The pressure solver module is also quite out of date, it hasn't been updated since topologies were introduced (#614) almost exactly a year ago.

This PR refactors the pressure solver module to:
1. Support all topologies on the CPU and GPU performing transposes and index permutations as needed by each transform.
2. Use the fastest transforms as allowed by the topology. This means batching dimensions when possible.
3. Consolidating all pressure solvers into a single solver for all topologies. This should simplify the code and make it easier to extend.

Resolves #586
Resolves #593
Resolves #594
Resolves #1007

# To batch or not to batch for FFTW on CPUs?

TODO:
- [x] Benchmark 1D {FFT, IFFT}{x, y, z}.
- [x] Benchmark 3D {FFT, IFFT}
- [x] Benchmark 1D {DCT, IDCT}{x, y, z}.
- [x] Benchmark 3D {DCT, IDCT}
- [x] Try N = 16, 64, 256
- [x] Is it faster to do 3 1D transforms or 1 3D transform? Answer: 1 3D transform.

To see whether we should just do 1D transforms for everything or whether batching is faster I ran some 1D and 3D FFT benchmarks. The results for triply-periodic are posted below.

Based on the benchmarks, it seems that for 256^3 doing three 1D transforms is ~15% slower than doing one 3D transform. So it makes sense to batch transforms when possible.

Note than FFT along dimension 1 is the fastest and FFT along dimension 2 is the slowest. FFT along dimension 3 is in the middle. So whatever FFTW is doing under the hood, FFTs along non-batched dimensions (e.g. along dimension 2) are slow.

```
                                               FFT benchmarks
┌───────────────┬─────┬───────────┬────────────┬────────────┬────────────┬────────────┬───────────┬────────┐
│ Architectures │  Ns │      dims │        min │     median │       mean │        max │    memory │ allocs │
├───────────────┼─────┼───────────┼────────────┼────────────┼────────────┼────────────┼───────────┼────────┤
│           CPU │  16 │ (1, 2, 3) │  13.948 μs │  14.043 μs │  20.717 μs │  80.605 μs │   0 bytes │      0 │
│           CPU │  64 │ (1, 2, 3) │   1.656 ms │   1.717 ms │   1.809 ms │   2.697 ms │   0 bytes │      0 │
│           CPU │ 256 │ (1, 2, 3) │ 229.619 ms │ 233.008 ms │ 234.033 ms │ 243.288 ms │   0 bytes │      0 │
│           CPU │  16 │         1 │   3.240 μs │   3.255 μs │   3.603 μs │   6.746 μs │   0 bytes │      0 │
│           CPU │  64 │         1 │ 445.803 μs │ 458.928 μs │ 513.041 μs │ 755.937 μs │   0 bytes │      0 │
│           CPU │ 256 │         1 │  61.083 ms │  63.464 ms │  63.969 ms │  67.009 ms │   0 bytes │      0 │
│           CPU │  16 │         2 │   4.085 μs │   4.135 μs │   4.723 μs │   8.088 μs │   0 bytes │      0 │
│           CPU │  64 │         2 │ 564.769 μs │ 579.278 μs │ 615.731 μs │ 804.859 μs │   0 bytes │      0 │
│           CPU │ 256 │         2 │ 110.718 ms │ 111.560 ms │ 111.506 ms │ 112.525 ms │   0 bytes │      0 │
│           CPU │  16 │         3 │   7.772 μs │   7.787 μs │   9.499 μs │  24.886 μs │   0 bytes │      0 │
│           CPU │  64 │         3 │ 684.541 μs │ 688.275 μs │ 811.874 μs │   1.463 ms │   0 bytes │      0 │
│           CPU │ 256 │         3 │  93.902 ms │  94.489 ms │  94.604 ms │  95.639 ms │   0 bytes │      0 │
└───────────────┴─────┴───────────┴────────────┴────────────┴────────────┴────────────┴───────────┴────────┘

3D FFT --> 3 × 1D FFTs slowdown:
CPU,  16: 1.0807x
CPU,  64: 1.0053x
CPU, 256: 1.1567x
```

# To batch or not to batch for CUFFT on GPUs?

We should investigate this separately for CUFFT since FFT along dimension 2 requires a transpose.

TODO:
- [x] Figure out how to do a FFT_y on the GPU!
- [x] Implement and benchmark doing it the distributed way.
- [x] Benchmark 1 3D FFT with 3 1D FFTs.
- [x] Benchmark 1 3D DCT with 3 1D DCTs.

Same benchmarks for the GPU are posted below. Batching is much faster (by a factor of 2-3) so we should batch when possible.

Note that FFTs along non-batched dimensions (dimension 2 in this case) are much slower since it involves two transpose operations.

Batching will not be possible for some topologies in which cases so we'll take a performance hit. But if the pressure solver is still 10~15% then a 2x hit on the pressure solver is not that large. The hit will mostly affect topologies we don't currently support anyways.

```
                                               FFT benchmarks
┌───────────────┬─────┬───────────┬────────────┬────────────┬────────────┬────────────┬───────────┬────────┐
│ Architectures │  Ns │      dims │        min │     median │       mean │        max │    memory │ allocs │
├───────────────┼─────┼───────────┼────────────┼────────────┼────────────┼────────────┼───────────┼────────┤
│           GPU │  16 │ (1, 2, 3) │  25.478 μs │  32.459 μs │ 122.062 μs │ 703.376 μs │ 224 bytes │     13 │
│           GPU │  64 │ (1, 2, 3) │  67.226 μs │  71.497 μs │ 146.042 μs │ 647.734 μs │ 224 bytes │     13 │
│           GPU │ 256 │ (1, 2, 3) │   2.982 ms │   3.041 ms │   3.036 ms │   3.116 ms │ 224 bytes │     13 │
│           GPU │  16 │         1 │  14.755 μs │  30.020 μs │ 107.932 μs │ 677.045 μs │  96 bytes │      5 │
│           GPU │  64 │         1 │  26.521 μs │  41.294 μs │ 114.587 μs │ 674.834 μs │  96 bytes │      5 │
│           GPU │ 256 │         1 │ 930.371 μs │ 936.222 μs │ 954.771 μs │   1.060 ms │  96 bytes │      5 │
│           GPU │  16 │         2 │  26.547 μs │  49.440 μs │ 127.426 μs │ 768.771 μs │  1.41 KiB │     59 │
│           GPU │  64 │         2 │ 116.160 μs │ 117.772 μs │ 193.909 μs │ 797.293 μs │  1.41 KiB │     59 │
│           GPU │ 256 │         2 │   4.963 ms │   5.010 ms │   5.014 ms │   5.073 ms │  1.41 KiB │     59 │
│           GPU │  16 │         3 │  14.918 μs │  22.509 μs │  40.029 μs │ 110.119 μs │ 224 bytes │     13 │
│           GPU │  64 │         3 │  40.151 μs │  45.495 μs │ 124.422 μs │ 646.093 μs │ 224 bytes │     13 │
│           GPU │ 256 │         3 │   1.062 ms │   1.067 ms │   1.101 ms │   1.292 ms │ 224 bytes │     13 │
└───────────────┴─────┴───────────┴────────────┴────────────┴────────────┴────────────┴───────────┴────────┘

3D FFT --> 3 × 1D FFTs slowdown:
GPU,  16: 3.1414x
GPU,  64: 2.8611x
GPU, 256: 2.3062x
```