@glwagner 2021-02-03
> Based on the benchmarks, it seems that for 256^3 doing three 1D transforms is ~15% slower than doing one 3D transform. So it makes sense to batch transforms when possible.

I think this makes sense given my primitive understanding of how FFTW picks optimal plans for the particular problem its asked to solve.