@ali-ramadhan 2021-02-04
Tests all pass but code is a bit messy (especially `plan_transforms.jl`) due to lots of special cases that I'm not yet sure how to simplify so needs some work.

Ran new FFT-based Poisson solver benchmarks on Tartarus (Titan V GPUs) and static ocean benchmarks for all topologies on Satori (Tesla V100 GPUs) + regression. Results are below. Will post a followup with some highlights/conclusions.

# FFT-based Poisson solver benchmarks

## Raw numbers

```
                                               FFT-based Poisson solver benchmarks
┌───────────────┬─────┬────────────────────────────────┬────────────┬────────────┬────────────┬────────────┬───────────┬────────┐
│ Architectures │  Ns │                     Topologies │        min │     median │       mean │        max │    memory │ allocs │
├───────────────┼─────┼────────────────────────────────┼────────────┼────────────┼────────────┼────────────┼───────────┼────────┤
│           CPU │ 192 │    (Bounded, Bounded, Bounded) │ 560.466 ms │ 563.145 ms │ 563.074 ms │ 566.700 ms │ 192 bytes │      4 │
│           CPU │ 192 │   (Bounded, Bounded, Periodic) │ 434.408 ms │ 435.974 ms │ 437.003 ms │ 441.246 ms │ 160 bytes │      2 │
│           CPU │ 192 │   (Bounded, Periodic, Bounded) │ 472.312 ms │ 473.340 ms │ 473.620 ms │ 475.649 ms │ 160 bytes │      2 │
│           CPU │ 192 │  (Bounded, Periodic, Periodic) │ 333.460 ms │ 334.702 ms │ 334.998 ms │ 336.918 ms │ 160 bytes │      2 │
│           CPU │ 192 │   (Periodic, Bounded, Bounded) │ 495.012 ms │ 497.853 ms │ 497.462 ms │ 500.181 ms │ 160 bytes │      2 │
│           CPU │ 192 │  (Periodic, Bounded, Periodic) │ 363.169 ms │ 365.104 ms │ 365.891 ms │ 373.893 ms │ 160 bytes │      2 │
│           CPU │ 192 │  (Periodic, Periodic, Bounded) │ 349.305 ms │ 350.431 ms │ 352.641 ms │ 371.861 ms │ 160 bytes │      2 │
│           CPU │ 192 │ (Periodic, Periodic, Periodic) │ 203.109 ms │ 203.653 ms │ 204.025 ms │ 206.834 ms │ 192 bytes │      4 │
│           GPU │ 192 │    (Bounded, Bounded, Bounded) │   7.765 ms │  16.841 ms │  15.934 ms │  16.872 ms │ 84.00 KiB │    904 │
│           GPU │ 192 │   (Bounded, Bounded, Periodic) │   6.492 ms │  13.599 ms │  12.878 ms │  13.633 ms │ 57.50 KiB │    651 │
│           GPU │ 192 │   (Bounded, Periodic, Bounded) │   6.432 ms │  13.616 ms │  12.883 ms │  13.640 ms │ 57.31 KiB │    645 │
│           GPU │ 192 │  (Bounded, Periodic, Periodic) │   9.430 ms │  19.467 ms │  18.452 ms │  19.582 ms │ 27.84 KiB │    294 │
│           GPU │ 192 │   (Periodic, Bounded, Bounded) │   6.330 ms │  13.532 ms │  12.824 ms │  13.642 ms │ 57.50 KiB │    651 │
│           GPU │ 192 │  (Periodic, Bounded, Periodic) │   4.882 ms │  10.317 ms │   9.772 ms │  10.332 ms │ 30.63 KiB │    386 │
│           GPU │ 192 │  (Periodic, Periodic, Bounded) │   3.424 ms │   7.083 ms │   6.713 ms │   7.175 ms │ 27.84 KiB │    294 │
│           GPU │ 192 │ (Periodic, Periodic, Periodic) │   1.843 ms │   3.482 ms │   3.318 ms │   3.489 ms │  1.09 KiB │     31 │
└───────────────┴─────┴────────────────────────────────┴────────────┴────────────┴────────────┴────────────┴───────────┴────────┘
```

## CPU to GPU speedup

```
             FFT-based Poisson solver CPU -> GPU speedup
┌─────┬────────────────────────────────┬─────────┬─────────┬────────┐
│  Ns │                     Topologies │ speedup │  memory │ allocs │
├─────┼────────────────────────────────┼─────────┼─────────┼────────┤
│ 192 │    (Bounded, Bounded, Bounded) │ 33.4393 │   448.0 │  226.0 │
│ 192 │   (Bounded, Bounded, Periodic) │ 32.0602 │   368.0 │  325.5 │
│ 192 │   (Bounded, Periodic, Bounded) │ 34.7631 │   366.8 │  322.5 │
│ 192 │  (Bounded, Periodic, Periodic) │ 17.1932 │   178.2 │  147.0 │
│ 192 │   (Periodic, Bounded, Bounded) │ 36.7915 │   368.0 │  325.5 │
│ 192 │  (Periodic, Bounded, Periodic) │ 35.3884 │   196.0 │  193.0 │
│ 192 │  (Periodic, Periodic, Bounded) │ 49.4769 │   178.2 │  147.0 │
│ 192 │ (Periodic, Periodic, Periodic) │ 58.4816 │ 5.83333 │   7.75 │
└─────┴────────────────────────────────┴─────────┴─────────┴────────┘
```

## CPU slowdown (vs. triply-periodic)

```
                  FFT-based Poisson solver relative performance (CPU)
┌───────────────┬─────┬────────────────────────────────┬──────────┬──────────┬────────┐
│ Architectures │  Ns │                     Topologies │ slowdown │   memory │ allocs │
├───────────────┼─────┼────────────────────────────────┼──────────┼──────────┼────────┤
│           CPU │ 192 │    (Bounded, Bounded, Bounded) │  2.76522 │      1.0 │    1.0 │
│           CPU │ 192 │   (Bounded, Bounded, Periodic) │  2.14077 │ 0.833333 │    0.5 │
│           CPU │ 192 │   (Bounded, Periodic, Bounded) │  2.32425 │ 0.833333 │    0.5 │
│           CPU │ 192 │  (Bounded, Periodic, Periodic) │  1.64349 │ 0.833333 │    0.5 │
│           CPU │ 192 │   (Periodic, Bounded, Bounded) │  2.44462 │ 0.833333 │    0.5 │
│           CPU │ 192 │  (Periodic, Bounded, Periodic) │  1.79278 │ 0.833333 │    0.5 │
│           CPU │ 192 │  (Periodic, Periodic, Bounded) │  1.72073 │ 0.833333 │    0.5 │
│           CPU │ 192 │ (Periodic, Periodic, Periodic) │      1.0 │      1.0 │    1.0 │
└───────────────┴─────┴────────────────────────────────┴──────────┴──────────┴────────┘
```

## GPU slowdown (vs. triply-periodic)

```
                  FFT-based Poisson solver relative performance (GPU)
┌───────────────┬─────┬────────────────────────────────┬──────────┬─────────┬─────────┐
│ Architectures │  Ns │                     Topologies │ slowdown │  memory │  allocs │
├───────────────┼─────┼────────────────────────────────┼──────────┼─────────┼─────────┤
│           GPU │ 192 │    (Bounded, Bounded, Bounded) │  4.83605 │    76.8 │ 29.1613 │
│           GPU │ 192 │   (Bounded, Bounded, Periodic) │  3.90501 │ 52.5714 │    21.0 │
│           GPU │ 192 │   (Bounded, Periodic, Bounded) │  3.91006 │    52.4 │ 20.8065 │
│           GPU │ 192 │  (Bounded, Periodic, Periodic) │  5.59024 │ 25.4571 │ 9.48387 │
│           GPU │ 192 │   (Periodic, Bounded, Bounded) │  3.88581 │ 52.5714 │    21.0 │
│           GPU │ 192 │  (Periodic, Bounded, Periodic) │  2.96267 │    28.0 │ 12.4516 │
│           GPU │ 192 │  (Periodic, Periodic, Bounded) │  2.03389 │ 25.4571 │ 9.48387 │
│           GPU │ 192 │ (Periodic, Periodic, Periodic) │      1.0 │     1.0 │     1.0 │
└───────────────┴─────┴────────────────────────────────┴──────────┴─────────┴─────────┘
```

# Static ocean benchmarks for all topologies

## Raw numbers

```
                                                    Topologies benchmarks
┌───────────────┬─────┬────────────────────────────────┬───────────┬───────────┬───────────┬───────────┬────────────┬────────┐
│ Architectures │  Ns │                     Topologies │       min │    median │      mean │       max │     memory │ allocs │
├───────────────┼─────┼────────────────────────────────┼───────────┼───────────┼───────────┼───────────┼────────────┼────────┤
│           CPU │ 192 │    (Bounded, Bounded, Bounded) │   2.402 s │   2.412 s │   2.413 s │   2.424 s │ 405.84 KiB │   2460 │
│           CPU │ 192 │   (Bounded, Bounded, Periodic) │   2.247 s │   2.250 s │   2.252 s │   2.259 s │ 363.28 KiB │   2162 │
│           CPU │ 192 │   (Bounded, Periodic, Bounded) │   1.890 s │   1.890 s │   1.890 s │   1.890 s │ 363.28 KiB │   2162 │
│           CPU │ 192 │  (Bounded, Periodic, Periodic) │   1.923 s │   1.933 s │   1.931 s │   1.936 s │ 317.00 KiB │   1806 │
│           CPU │ 192 │   (Periodic, Bounded, Bounded) │   1.864 s │   1.869 s │   1.868 s │   1.871 s │ 363.28 KiB │   2162 │
│           CPU │ 192 │  (Periodic, Bounded, Periodic) │   1.685 s │   1.686 s │   1.688 s │   1.693 s │ 317.00 KiB │   1806 │
│           CPU │ 192 │  (Periodic, Periodic, Bounded) │   2.092 s │   2.114 s │   2.109 s │   2.121 s │ 317.00 KiB │   1806 │
│           CPU │ 192 │ (Periodic, Periodic, Periodic) │   1.780 s │   1.796 s │   1.801 s │   1.828 s │ 277.47 KiB │   1662 │
│           GPU │ 192 │    (Bounded, Bounded, Bounded) │ 14.888 ms │ 22.339 ms │ 21.623 ms │ 22.605 ms │ 913.08 KiB │   9570 │
│           GPU │ 192 │   (Bounded, Bounded, Periodic) │ 13.213 ms │ 18.322 ms │ 17.794 ms │ 18.914 ms │ 927.17 KiB │   8415 │
│           GPU │ 192 │   (Bounded, Periodic, Bounded) │ 12.658 ms │ 18.326 ms │ 17.780 ms │ 18.495 ms │ 930.73 KiB │   8289 │
│           GPU │ 192 │  (Bounded, Periodic, Periodic) │ 15.206 ms │ 22.759 ms │ 21.995 ms │ 22.799 ms │ 935.36 KiB │   6976 │
│           GPU │ 192 │   (Periodic, Bounded, Bounded) │ 12.717 ms │ 18.315 ms │ 17.841 ms │ 18.997 ms │ 930.92 KiB │   8295 │
│           GPU │ 192 │  (Periodic, Bounded, Periodic) │ 12.404 ms │ 15.545 ms │ 15.266 ms │ 15.972 ms │ 938.14 KiB │   7068 │
│           GPU │ 192 │  (Periodic, Periodic, Bounded) │ 10.097 ms │ 13.083 ms │ 12.793 ms │ 13.159 ms │ 939.77 KiB │   6898 │
│           GPU │ 192 │ (Periodic, Periodic, Periodic) │  8.948 ms │ 10.050 ms │  9.948 ms │ 10.128 ms │ 945.39 KiB │   5625 │
└───────────────┴─────┴────────────────────────────────┴───────────┴───────────┴───────────┴───────────┴────────────┴────────┘
```

## CPU to GPU speedup

```
                    Topologies CPU -> GPU speedup
┌─────┬────────────────────────────────┬─────────┬─────────┬─────────┐
│  Ns │                     Topologies │ speedup │  memory │  allocs │
├─────┼────────────────────────────────┼─────────┼─────────┼─────────┤
│ 192 │    (Bounded, Bounded, Bounded) │ 107.967 │ 2.24983 │ 3.89024 │
│ 192 │   (Bounded, Bounded, Periodic) │ 122.789 │ 2.55222 │ 3.89223 │
│ 192 │   (Bounded, Periodic, Bounded) │ 103.133 │ 2.56202 │ 3.83395 │
│ 192 │  (Bounded, Periodic, Periodic) │  84.934 │ 2.95066 │ 3.86268 │
│ 192 │   (Periodic, Bounded, Bounded) │  102.06 │ 2.56254 │ 3.83673 │
│ 192 │  (Periodic, Bounded, Periodic) │ 108.457 │ 2.95943 │ 3.91362 │
│ 192 │  (Periodic, Periodic, Bounded) │ 161.616 │ 2.96456 │ 3.81949 │
│ 192 │ (Periodic, Periodic, Periodic) │ 178.682 │  3.4072 │ 3.38448 │
└─────┴────────────────────────────────┴─────────┴─────────┴─────────┘
```

## CPU slowdown (vs. triply-periodic)

```
                         Topologies relative performance (CPU)
┌───────────────┬─────┬────────────────────────────────┬──────────┬─────────┬─────────┐
│ Architectures │  Ns │                     Topologies │ slowdown │  memory │  allocs │
├───────────────┼─────┼────────────────────────────────┼──────────┼─────────┼─────────┤
│           CPU │ 192 │    (Bounded, Bounded, Bounded) │  1.34309 │ 1.46266 │ 1.48014 │
│           CPU │ 192 │   (Bounded, Bounded, Periodic) │  1.25281 │ 1.30927 │ 1.30084 │
│           CPU │ 192 │   (Bounded, Periodic, Bounded) │  1.05249 │ 1.30927 │ 1.30084 │
│           CPU │ 192 │  (Bounded, Periodic, Periodic) │  1.07645 │ 1.14247 │ 1.08664 │
│           CPU │ 192 │   (Periodic, Bounded, Bounded) │   1.0409 │ 1.30927 │ 1.30084 │
│           CPU │ 192 │  (Periodic, Bounded, Periodic) │ 0.938853 │ 1.14247 │ 1.08664 │
│           CPU │ 192 │  (Periodic, Periodic, Bounded) │  1.17749 │ 1.14247 │ 1.08664 │
│           CPU │ 192 │ (Periodic, Periodic, Periodic) │      1.0 │     1.0 │     1.0 │
└───────────────┴─────┴────────────────────────────────┴──────────┴─────────┴─────────┘
```

## GPU slowdown (vs. triply-periodic)

```
                         Topologies relative performance (GPU)
┌───────────────┬─────┬────────────────────────────────┬──────────┬──────────┬─────────┐
│ Architectures │  Ns │                     Topologies │ slowdown │   memory │  allocs │
├───────────────┼─────┼────────────────────────────────┼──────────┼──────────┼─────────┤
│           GPU │ 192 │    (Bounded, Bounded, Bounded) │  2.22277 │ 0.965821 │ 1.70133 │
│           GPU │ 192 │   (Bounded, Bounded, Periodic) │  1.82308 │ 0.980729 │   1.496 │
│           GPU │ 192 │   (Bounded, Periodic, Bounded) │  1.82349 │ 0.984497 │  1.4736 │
│           GPU │ 192 │  (Bounded, Periodic, Periodic) │  2.26462 │ 0.989389 │ 1.24018 │
│           GPU │ 192 │   (Periodic, Bounded, Bounded) │  1.82237 │ 0.984695 │ 1.47467 │
│           GPU │ 192 │  (Periodic, Bounded, Periodic) │  1.54676 │ 0.992331 │ 1.25653 │
│           GPU │ 192 │  (Periodic, Periodic, Bounded) │  1.30183 │  0.99405 │ 1.22631 │
│           GPU │ 192 │ (Periodic, Periodic, Periodic) │      1.0 │      1.0 │     1.0 │
└───────────────┴─────┴────────────────────────────────┴──────────┴──────────┴─────────┘
```

# Performance vs. main branch

## Main branch

```
                                                    Topologies benchmarks
┌───────────────┬─────┬────────────────────────────────┬───────────┬───────────┬───────────┬───────────┬────────────┬────────┐
│ Architectures │  Ns │                     Topologies │       min │    median │      mean │       max │     memory │ allocs │
├───────────────┼─────┼────────────────────────────────┼───────────┼───────────┼───────────┼───────────┼────────────┼────────┤
│           CPU │ 192 │   (Periodic, Bounded, Bounded) │   1.922 s │   1.922 s │   1.967 s │   2.058 s │ 363.61 KiB │   2163 │
│           CPU │ 192 │  (Periodic, Periodic, Bounded) │   2.143 s │   2.144 s │   2.145 s │   2.146 s │ 317.33 KiB │   1807 │
│           CPU │ 192 │ (Periodic, Periodic, Periodic) │   1.791 s │   1.793 s │   1.793 s │   1.794 s │ 277.77 KiB │   1661 │
│           GPU │ 192 │   (Periodic, Bounded, Bounded) │ 32.188 ms │ 37.447 ms │ 36.936 ms │ 37.557 ms │ 985.94 KiB │  13476 │
│           GPU │ 192 │  (Periodic, Periodic, Bounded) │ 11.051 ms │ 11.114 ms │ 11.148 ms │ 11.533 ms │ 807.44 KiB │  10746 │
│           GPU │ 192 │ (Periodic, Periodic, Periodic) │  9.859 ms │ 10.104 ms │ 10.136 ms │ 10.682 ms │ 707.81 KiB │   9469 │
└───────────────┴─────┴────────────────────────────────┴───────────┴───────────┴───────────┴───────────┴────────────┴────────┘
```

## This PR/branch

```
                                                    Topologies benchmarks
┌───────────────┬─────┬────────────────────────────────┬───────────┬───────────┬───────────┬───────────┬────────────┬────────┐
│ Architectures │  Ns │                     Topologies │       min │    median │      mean │       max │     memory │ allocs │
├───────────────┼─────┼────────────────────────────────┼───────────┼───────────┼───────────┼───────────┼────────────┼────────┤
│           CPU │ 192 │   (Periodic, Bounded, Bounded) │   1.864 s │   1.869 s │   1.868 s │   1.871 s │ 363.28 KiB │   2162 │
│           CPU │ 192 │  (Periodic, Periodic, Bounded) │   2.092 s │   2.114 s │   2.109 s │   2.121 s │ 317.00 KiB │   1806 │
│           CPU │ 192 │ (Periodic, Periodic, Periodic) │   1.780 s │   1.796 s │   1.801 s │   1.828 s │ 277.47 KiB │   1662 │
│           GPU │ 192 │   (Periodic, Bounded, Bounded) │ 12.717 ms │ 18.315 ms │ 17.841 ms │ 18.997 ms │ 930.92 KiB │   8295 │
│           GPU │ 192 │  (Periodic, Periodic, Bounded) │ 10.097 ms │ 13.083 ms │ 12.793 ms │ 13.159 ms │ 939.77 KiB │   6898 │
│           GPU │ 192 │ (Periodic, Periodic, Periodic) │  8.948 ms │ 10.050 ms │  9.948 ms │ 10.128 ms │ 945.39 KiB │   5625 │
└───────────────┴─────┴────────────────────────────────┴───────────┴───────────┴───────────┴───────────┴────────────┴────────┘
```